+++ src/Symfony/Component/Scheduler/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add a "Next Run In" column to `debug:scheduler` showing the time until the next run
  * Add `env` option to `#[AsCronTask]` and `#[AsPeriodicTask]` to restrict a task to one or more environments
+ * Deprecate `Schedule::with()`, use `add()` on a new `Schedule` instead
 
 8.1
 ---